Tests: Allow reading selection of sample logs
Sample logs are short and only from publicly-logged channels. When I
finally bring pytest into the project, I'll be using the sample logs.

The CLI flag "--log-dir" allows specifying a directory from which to
read logs; this can be used for testing.

Split main() into two functions; one for generating the table, the other
for prettifying and printing it. The former will be used for testing.